-Modify ByteArray.toASCII()

+++ src/jsMain/kotlin/Utility.kt
@@ -24,7 +24,7 @@ class Utility {
             val asciiChars = CharArray(this.size)
 
             for (i in indices) {
-                asciiChars[i] = this[i].toChar()
+                asciiChars[i] = this[i].toInt().toChar()
             }
 
             return asciiChars.concatToString()
